se agregan atributos en reserva y usuario

+++ controllers/reservaController.js
@@ -1,11 +1,10 @@
 const Reserva = require('../models/Reserva.js');
 const Estado = require('../models/Estado.js');
-const Usuario = require('../models/Usuario.js');
-const {sendEmailState} = require('../services/emailService.js');
+const serviceUpdate = require('../services/updateReservaEstado.js');
 
 const getReservas = async (req, res) => {
     try {
-        const reservas = await Reserva.find().populate('usuario', 'nombre').populate('mascotas', 'nombre').populate('estado', 'estado');
+        const reservas = await Reserva.find().populate('cliente', 'nombre').populate('cuidador', 'nombre').populate('mascotas', 'nombre').populate('estado', 'estado');
         res.status(200).json(reservas);
     } catch (error) {
         res.status(400).json({ message: error.message });
@@ -14,9 +13,9 @@ const getReservas = async (req, res) => {
 
 const createReserva = async (req, res) => {
     try {
-        const { fechaInicio, fechaFin, comentario, tarifaTurno, usuario, mascotas} = req.body;
-        const estado = await Estado.findOne({ estado: 'Pendiente' });;
-        const newReserva = await Reserva.create({ fechaInicio, fechaFin, comentario, tarifaTurno, usuario, mascotas, estado: estado._id});
+        const { fechaInicio, fechaFin, comentario, tarifaTurno, cliente, cuidador, mascotas} = req.body;
+        const estado = await Estado.findOne({ estado: 'Pendiente' });
+        const newReserva = await Reserva.create({ fechaInicio, fechaFin, comentario, tarifaTurno, cliente, cuidador, mascotas, estado: estado._id});
         res.status(201).json(newReserva);
     } catch (error) {
         console.log(error.message);
@@ -27,6 +26,11 @@ const createReserva = async (req, res) => {
 const deleteReserva = async (req, res) => {
     try {
         const idReserva = req.params.id;
+        const reserva = await Reserva.findById(idReserva);
+        const estado = await Estado.findOne({ estado: 'Pendiente' });
+        if (reserva.estado != estado._id) {
+            return res.status(400).json({ message: 'No se puede eliminar una reserva que no esté en estado Pendiente' });
+        }
         const result = await Reserva.deleteOne({_id:idReserva});
         res.status(200).json(result); 
     } catch (error) {
@@ -45,10 +49,20 @@ const editReserva = async (req, res) => {
     } 
 }
 
-const getReservasPorUsuario = async (req, res) => {
+const getReservasPorCliente = async (req, res) => {
     try {
-        const idUsuario = req.params.id;
-        const reservas = await Reserva.find({usuario: idUsuario}).populate('usuario', 'nombre').populate('mascotas', 'nombre').populate('estado', 'estado');
+        const idCliente = req.params.id;
+        const reservas = await Reserva.find({cliente: idCliente}).populate('cuidador', 'nombre').populate('mascotas', 'nombre').populate('estado', 'estado');
+        res.status(200).json(reservas);
+    } catch (error) {
+        res.status(400).json({ message: error.message });
+    }
+}
+
+const getReservasPorCuidador = async (req, res) => {
+    try {
+        const idCuidador = req.params.id;
+        const reservas = await Reserva.find({cuidador: idCuidador}).populate('cliente', 'nombre').populate('mascotas', 'nombre').populate('estado', 'estado');
         res.status(200).json(reservas);
     } catch (error) {
         res.status(400).json({ message: error.message });
@@ -58,7 +72,7 @@ const getReservasPorUsuario = async (req, res) => {
 const getOneReserva = async (req, res) => {
     try {
         const idReserva = req.params.id;
-        const reserva = await Reserva.findById(idReserva).populate('usuario', 'nombre').populate('mascotas', 'nombre').populate('estado', 'estado');
+        const reserva = await Reserva.findById(idReserva).populate('cliente', 'nombre').populate('cuidador', 'nombre').populate('mascotas', 'nombre').populate('estado', 'estado');
         res.status(200).json(reserva);
     } catch (error) {
         res.status(400).json({ message: error.message });
@@ -69,21 +83,7 @@ const updateReservaEstado = async (req, res) => {
     try {
         const idReserva = req.params.idReserva;
         const idEstado = req.params.idEstado;
-        const reserva = await Reserva.findById(idReserva);
-        if (!reserva) {
-            return res.status(404).json({ message: 'Reserva no encontrada' });
-        }
-        const estadoNuevo = await Estado.findById(idEstado);
-        if (!estadoNuevo) {
-            return res.status(404).json({ message: 'Estado no encontrado' });
-        }
-        const usuario = await Usuario.findById(reserva.usuario);
-        if (!usuario) {
-            return res.status(404).json({ message: 'Usuario no encontrado' });
-        }
-        reserva.estado = estadoNuevo._id;
-        await reserva.save();
-        sendEmailState(usuario.email, estadoNuevo.estado);
+        const reserva = await serviceUpdate.updateReservaEstado(idReserva, idEstado);
         res.status(200).json(reserva);
     } catch (error) {
         res.status(400).json({ message: error.message });
@@ -95,7 +95,8 @@ module.exports = {
     createReserva,
     deleteReserva,
     editReserva,
-    getReservasPorUsuario,
+    getReservasPorCliente,
+    getReservasPorCuidador,
     getOneReserva,
     updateReservaEstado
 }
